Add files via upload
This Graph represents how the amount of work we can get changes with the difference in internal energies. I am going to make an assumption based around the center of the graph initially and figure out the next states and hopefully get an initial efficiency number.  Also, This graph is using the equation dQ = dU+dW I am assuming that dQ is zero due to this process being isentropic and ending up with dW = -dU.